+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This file documents all notable changes made up to version 2.0.0 for each projec
 - Add Ibanez 6 string template.
 - Add selector for handed template mode.
 
+## [2.3.0] - 2025-10-20
+- Add Ibanez 6 string template.
+- Add selector for handed template mode.
+
 ## [2.2.2] - 2025-09-27
 - Update tunning for 4 strings.